Moving toward enabling DI of the Button.
